refactor(gateway): extract DaoDepositReader (#106 phase 3)
Pulls the ~210-line getDaoDeposits inline body out of GatewayRepository.
GatewayRepository drops from 2292 to 2087 lines.

DaoDepositReader.list(walletScript, currentEpoch, network):
- Cursor-paginated lock-script cell scan via nativeGetCells
- Local filter on DAO type-script code hash
- Cursor-paginated tx scan for spent outpoints, filter to live cells
- Per-cell header resolution via DaoHeaderResolver (composes phase 2)
- Per-cell compensation math: deposit vs withdrawing branches,
  max-withdraw + unlock-epoch JNI calls
- APC + cycle-progress derivation, status assignment

GatewayRepository.getDaoDeposits keeps the public Result<List<DaoDeposit>>
shape — body now four lines that compose currentEpoch + the reader.

depositToDao / withdrawFromDao / unlockDao stay on GatewayRepository
because they orchestrate the send pipeline (cell selection,
TransactionBuilder, sendTransaction). Pulling them out would need to
thread the entire send-side state, which is the natural boundary for
a future SendCoordinator extraction.

Cumulative #106 progress: 2629 → 2087 lines (~21% reduction). 611/611
tests pass. Debug build installs + cold-launches cleanly with the new
DI wiring.

Refs #106
