+++ android/app/src/main/java/com/rjnr/pocketnode/data/gateway/GatewayRepository.kt
@@ -114,6 +114,7 @@ class GatewayRepository @Inject constructor(
     private val broadcastClient: BroadcastClient,
     private val syncCoordinator: SyncCoordinator,
     private val daoHeaderResolver: DaoHeaderResolver,
+    private val daoDepositReader: DaoDepositReader,
 ) : TipSource {
     private val sendMutex = Mutex()
 
@@ -1740,216 +1741,10 @@ class GatewayRepository @Inject constructor(
 
     suspend fun getDaoDeposits(): Result<List<DaoDeposit>> = runCatching {
         val info = _walletInfo.value ?: throw Exception("No wallet")
-
-        // Query ALL cells by lock script (like Neuron), then filter locally for DAO type
-        val searchKey = JniSearchKey(
-            script = info.script,
-            scriptType = "lock",
-            withData = true
-        )
-        val searchKeyJson = json.encodeToString(searchKey)
-
-        // Paginate all cells by lock script
-        val allCellObjects = mutableListOf<JniCell>()
-        var cellsCursor: String? = null
-        do {
-            val pageJson = LightClientNative.nativeGetCells(searchKeyJson, "desc", 100, cellsCursor)
-                ?: break
-            val page = json.decodeFromString<JniPagination<JniCell>>(pageJson)
-            allCellObjects += page.objects
-            cellsCursor = page.lastCursor?.takeUnless { it == cellsCursor }
-        } while (cellsCursor != null)
-
-        // Filter locally: only cells whose type script matches DAO code hash
-        val daoCells = allCellObjects.filter { cell ->
-            cell.output.type?.codeHash == DaoConstants.DAO_CODE_HASH
-        }
-        Log.d(TAG, "📋 getDaoDeposits: ${daoCells.size} DAO cells out of ${allCellObjects.size} total")
-
-        // Paginate transactions to find spent outpoints
-        val spentOutpoints = mutableSetOf<String>()
-        var txCursor: String? = null
-        do {
-            val txJson = LightClientNative.nativeGetTransactions(searchKeyJson, "desc", 100, txCursor)
-                ?: break
-            val txPag = json.decodeFromString<JniPagination<JniTxWithCell>>(txJson)
-            txPag.objects.forEach { txWithCell ->
-                txWithCell.transaction.inputs.forEach { input ->
-                    spentOutpoints.add("${input.previousOutput.txHash}:${input.previousOutput.index}")
-                }
-            }
-            txCursor = txPag.lastCursor?.takeUnless { it == txCursor }
-        } while (txCursor != null)
-
-        val liveDaoCells = daoCells.filter { cell ->
-            val key = "${cell.outPoint.txHash}:${cell.outPoint.index}"
-            key !in spentOutpoints
-        }
-        Log.d(TAG, "📋 getDaoDeposits: ${liveDaoCells.size} live DAO cells")
-
-        val pagination = JniPagination(objects = liveDaoCells, lastCursor = null)
-
         val currentEpoch = getCurrentEpoch().getOrNull()
-
-        val deposits = mutableListOf<DaoDeposit>()
-        for (jniCell in pagination.objects) {
-            val cell = jniCell.toCell()
-            val data = cell.data.removePrefix("0x")
-            val cellId = "${cell.outPoint.txHash.take(20)}...:${cell.outPoint.index}"
-            val capacityShannons = cell.capacity.removePrefix("0x").toLong(16)
-
-            // Determine if deposit or withdrawing cell
-            val isWithdrawing = data.length == 16 && data != "0000000000000000"
-            Log.d(TAG, "🏦 Processing DAO cell: $cellId, data=$data, isWithdrawing=$isWithdrawing, capacity=$capacityShannons")
-
-            var depositBlockNumber = cell.blockNumber.removePrefix("0x").toLong(16)
-            var depositBlockHash = "" // may remain empty if header unavailable
-            var depositEpoch: EpochInfo? = null
-            var withdrawBlockNumber: Long? = null
-            var withdrawBlockHash: String? = null
-            var withdrawEpoch: EpochInfo? = null
-            var compensation = 0L
-            var unlockEpoch: EpochInfo? = null
-            var lockRemainingHours: Int? = null
-            var depositTimestampMs = 0L
-            var apc = 0.0
-
-            // Get block header for compensation & epoch data (cache-first).
-            val blockHash = getBlockHashForCell(cell.outPoint.txHash)
-            val cellBlockHeader = if (blockHash != null) {
-                getOrFetchHeader(blockHash)
-            } else null
-
-            if (cellBlockHeader != null) {
-                val cellBlockEpoch = EpochInfo.fromHex(cellBlockHeader.epoch)
-                depositBlockHash = cellBlockHeader.hash
-                depositEpoch = cellBlockEpoch
-                // Default timestamp from this cell's block; overridden for withdrawing cells below
-                depositTimestampMs = cellBlockHeader.timestamp.removePrefix("0x").toLong(16)
-
-                if (isWithdrawing) {
-                    // Cell data contains deposit block number as 8-byte LE
-                    val depositBlockNum = data.chunked(2)
-                        .map { it.toInt(16).toByte() }
-                        .toByteArray()
-                        .let { bytes ->
-                            var num = 0L
-                            for (i in bytes.indices) {
-                                num = num or ((bytes[i].toLong() and 0xFF) shl (i * 8))
-                            }
-                            num
-                        }
-
-                    withdrawBlockNumber = cell.blockNumber.removePrefix("0x").toLong(16)
-                    withdrawBlockHash = cellBlockHeader.hash
-                    withdrawEpoch = cellBlockEpoch
-                    depositBlockNumber = depositBlockNum
-
-                    // Get original deposit header for compensation (cache-first)
-                    val withdrawTxJson = LightClientNative.nativeGetTransaction(cell.outPoint.txHash)
-                    val withdrawTx = withdrawTxJson?.let { json.decodeFromString<JniTransactionWithStatus>(it) }
-                    val origDepositBlockHash = withdrawTx?.transaction?.headerDeps?.firstOrNull()
-                    // Preserve the hash even if full header fetch fails
-                    if (origDepositBlockHash != null) {
-                        depositBlockHash = origDepositBlockHash
-                    }
-                    val origDepositHeader = origDepositBlockHash?.let { h ->
-                        getOrFetchHeader(h)
-                    }
-
-                    if (origDepositHeader != null) {
-                        depositBlockHash = origDepositHeader.hash
-                        depositEpoch = EpochInfo.fromHex(origDepositHeader.epoch)
-                        depositTimestampMs = origDepositHeader.timestamp.removePrefix("0x").toLong(16)
-                        val maxWithdraw = LightClientNative.nativeCalculateMaxWithdraw(
-                            origDepositHeader.dao, cellBlockHeader.dao, capacityShannons, 61_00000000L
-                        )
-                        if (maxWithdraw >= 0) compensation = maxWithdraw - capacityShannons
-                        val sinceHex = LightClientNative.nativeCalculateUnlockEpoch(
-                            origDepositHeader.epoch, cellBlockHeader.epoch
-                        )
-                        if (sinceHex != null) {
-                            val epochVal = sinceHex.removePrefix("0x").toLong(16) and 0x00FF_FFFF_FFFF_FFFFL
-                            unlockEpoch = EpochInfo.fromHex("0x${epochVal.toString(16)}")
-                            if (currentEpoch != null && unlockEpoch!!.value > currentEpoch.value) {
-                                lockRemainingHours = ((unlockEpoch!!.value - currentEpoch.value) * DaoConstants.HOURS_PER_EPOCH).toInt()
-                            }
-                        }
-                    }
-                } else {
-                    // Deposited cell — calculate compensation using current tip header
-                    val tipJson = LightClientNative.nativeGetTipHeader()
-                    if (tipJson != null) {
-                        val tipHeader = json.decodeFromString<JniHeaderView>(tipJson)
-                        val maxWithdraw = LightClientNative.nativeCalculateMaxWithdraw(
-                            cellBlockHeader.dao, tipHeader.dao, capacityShannons, 61_00000000L
-                        )
-                        if (maxWithdraw >= 0) compensation = maxWithdraw - capacityShannons
-                    }
-                }
-            } else {
-                Log.d(TAG, "  No header available for $cellId — showing deposit with basic info")
-                if (isWithdrawing) {
-                    val depositBlockNum = data.chunked(2)
-                        .map { it.toInt(16).toByte() }
-                        .toByteArray()
-                        .let { bytes ->
-                            var num = 0L
-                            for (i in bytes.indices) {
-                                num = num or ((bytes[i].toLong() and 0xFF) shl (i * 8))
-                            }
-                            num
-                        }
-                    withdrawBlockNumber = cell.blockNumber.removePrefix("0x").toLong(16)
-                    depositBlockNumber = depositBlockNum
-                }
-            }
-
-            // Compute per-deposit APC when enough time has elapsed
-            if (compensation > 0 && depositTimestampMs > 0 && capacityShannons > 0) {
-                val elapsedDays = (System.currentTimeMillis() - depositTimestampMs) / 86_400_000.0
-                if (elapsedDays >= 1.0) {
-                    apc = (compensation.toDouble() / capacityShannons) / (elapsedDays / 365.25) * 100.0
-                }
-            }
-
-            // Calculate cycle progress (best-effort with available epoch data)
-            val depositedEpochs = if (currentEpoch != null && depositEpoch != null) {
-                (currentEpoch.value - depositEpoch.value).coerceAtLeast(0.0)
-            } else 0.0
-            val cycleProgress = ((depositedEpochs % DaoConstants.WITHDRAW_EPOCHS) / DaoConstants.WITHDRAW_EPOCHS).toFloat()
-
-            val status = determineDaoStatus(
-                isWithdrawingCell = isWithdrawing,
-                hasPendingWithdraw = false,
-                hasPendingUnlock = false,
-                hasPendingDeposit = false,
-                currentEpoch = currentEpoch,
-                unlockEpoch = unlockEpoch
-            )
-
-            deposits.add(DaoDeposit(
-                outPoint = cell.outPoint,
-                capacity = capacityShannons,
-                status = status,
-                depositBlockNumber = depositBlockNumber,
-                depositBlockHash = depositBlockHash,
-                depositEpoch = depositEpoch,
-                withdrawBlockNumber = withdrawBlockNumber,
-                withdrawBlockHash = withdrawBlockHash,
-                withdrawEpoch = withdrawEpoch,
-                compensation = compensation.coerceAtLeast(0L),
-                unlockEpoch = unlockEpoch,
-                lockRemainingHours = lockRemainingHours,
-                compensationCycleProgress = cycleProgress,
-                cyclePhase = cyclePhaseFromProgress(cycleProgress),
-                depositTimestamp = depositTimestampMs,
-                apc = apc
-            ))
-            Log.d(TAG, "✅ DAO deposit added: $cellId, status=$status, capacity=$capacityShannons")
-        }
-        deposits
+        daoDepositReader.list(info.script, currentEpoch, currentNetwork)
     }
+
 
     suspend fun getDaoOverview(): Result<DaoOverview> = runCatching {
         val deposits = getDaoDeposits().getOrThrow()
